+++ .github/workflows/firebase-functions-merge.yml
@@ -1,10 +1,11 @@
 name: Deploy Firebase Functions on merge
-'on':
+on:
   push:
     branches:
       - main
     paths:
       - "functions/**"
+      - "shared/**"
 jobs:
   build_and_deploy:
     runs-on: ubuntu-latest
