+++ .github/workflows/firebase-hosting-merge.yml
@@ -1,10 +1,11 @@
 name: Deploy to Firebase Hosting on merge
-'on':
+on:
   push:
     branches:
       - main
     paths:
       - "client/**"
+      - "shared/**"
 jobs:
   build_and_deploy:
     runs-on: ubuntu-latest
