+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Gunn-WATT",
+  "name": "@watt/client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
