+++ client/src/components/lists/ClubComponent.tsx
@@ -1,12 +1,7 @@
 import { useState } from 'react';
 import ClubComponentModal from './ClubComponentModal';
+import {Club} from '@watt/shared/data/clubs';
 
-
-export type Club = {
-    name: string, new?: boolean, room: string, desc: string, day: string, time: string,
-    zoom?: string, video?: string, signup?: string, tier: number,
-    prez: string, advisor: string, email: string, coadvisor?: string, coemail?: string;
-}
 
 export default function ClubComponent(props: Club & {id: string}) {
     const {name, room, day} = props;
