+++ shared/data/staff.ts
@@ -1,4 +1,20 @@
-import {Staff} from '../components/lists/StaffComponent';
+/*
+The period data structure is a bunch of nested Objects, where each period is represented by its name
+(1, 2, 3, SELF, etc.) and contains data structured like so:
+{1: [Class, Room], 2: [Class, Room]}
+where 1 and 2 represent first and second semester. However, we also need to support when two classes
+are taught at once in one period, where the data would be structured instead like
+{1: {1: [Class, Room], 2: [Class, Room]} 2: ...}
+*/
+
+export type SemesterClassObj = [string, string | null] | 'none';
+export type ClassObj = SemesterClassObj | {1: SemesterClassObj, 2: SemesterClassObj};
+export type PeriodObj = {1: ClassObj, 2: ClassObj};
+export type Staff = {
+    name: string, title?: string, email?: string, room?: string,
+    dept?: string, phone?: string, periods?: {[key: string]: PeriodObj},
+    other?: string // "other" info like "Teaches SELF", "Has Counseling"
+};
 
 const data: {timestamp: string, data: {[key: string]: Staff}} = {
     "timestamp": "2021-09-19T01:59:48.232Z",
