+++ shared/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@watt/shared",
+  "version": "1.0.0",
+  "description": "Shared utilities for client, functions, and scripts.",
+  "dependencies": {
+    "luxon": "^3.0.1"
+  },
+  "devDependencies": {
+    "typescript": "^4.7.4"
+  }
+}
